+++ src/adsctl/cli/auth.py
@@ -8,6 +8,8 @@ from urllib.parse import unquote
 import click
 from google_auth_oauthlib.flow import Flow
 
+from adsctl.application import Application
+
 _SCOPE = "https://www.googleapis.com/auth/adwords"
 _SERVER = "127.0.0.1"
 _PORT = 8080
@@ -19,16 +21,87 @@ _REDIRECT_URI = f"http://{_SERVER}:{_PORT}"
     "secrets_file",
     type=click.Path(exists=True),
 )
-@click.pass_context
-def auth(ctx: click.Context, secrets_file):
+@click.pass_obj
+def auth(app: Application, secrets_file):
     """Authenticate with Google Ads API."""
 
-    # Do the OAuth flow to get a refresh token.
+    # Load config
+    app.load_config()
 
-    configured_scopes = [_SCOPE]
+    current_account = app.config.current_account
+    click.echo("Getting new refresh token for account: " + current_account)
+
+    new_refresh_token = oauth_flow(secrets_file)
+
+    if new_refresh_token is None:
+        click.echo("Authentication failed.")
+        sys.exit(1)
+
+    click.echo(f"Your new refresh token is: {new_refresh_token}\n")
+    click.echo("Updating config file to include refresh token.")
+
+    with open(app.config_file.path, "r") as f:
+        current_contents = f.read()
+
+    with open(app.config_file.path, "w") as f:
+        new_content = current_contents
+        current_token = app.config.accounts[current_account].oauth.refresh_token
+
+        # None means the refresh_token field not found in the account header
+        if current_token is None:
+            # Add the refresh token to the config file below the account.oauth header
+
+            account_header_str = f"[accounts.{current_account}.oauth]"
+
+            header_occurrences = current_contents.count(account_header_str)
+
+            if header_occurrences == 1:
+                # Found the right header, so append the refresh token
+                appended_line = (
+                    account_header_str + f'\nrefresh_token = "{new_refresh_token}"'
+                )
+                new_content = new_content.replace(account_header_str, appended_line)
+            else:
+                f.write(current_contents)
+                click.echo(
+                    "Error updating the config file:\n"
+                    "Could't find the right account group."
+                    "Please add the refresh token above to the config file:\n"
+                    f"{app.config_file.path}"
+                )
+                sys.exit(1)
+        else:
+            # See how many matches of the refresh_token find in the file
+
+            current_token_str = f'refresh_token = "{current_token}"'
+            current_token_occurrences = current_contents.count(current_token_str)
+
+            if current_token_occurrences == 1:
+                # We found one match, so replace it
+                new_token_str = rf'refresh_token = "{new_refresh_token}"'
+                new_content = current_contents.replace(current_token_str, new_token_str)
+            else:
+                f.write(current_contents)
+                click.echo(
+                    "Error updating the config file:\n"
+                    "The current refresh_token was found multiple times in the file. "
+                    "Please add the refresh token above to the config file:\n"
+                    f"{app.config_file.path}"
+                )
+                sys.exit(1)
+
+        f.write(new_content)
+
+    click.echo("Done! You can now run adsctl commands.")
+
+
+def oauth_flow(secrets_file: str) -> str | None:
+    """Do the OAuth flow to get a refresh token."""
 
     # if args.additional_scopes:
     #     configured_scopes.extend(args.additional_scopes)
+
+    configured_scopes = [_SCOPE]
 
     flow = Flow.from_client_secrets_file(secrets_file, scopes=configured_scopes)
     flow.redirect_uri = _REDIRECT_URI
@@ -48,9 +121,9 @@ def auth(ctx: click.Context, secrets_file):
     # typical web application you would redirect the user to this URL, and they
     # would be redirected back to "redirect_url" provided earlier after
     # granting permission.
-    print("Paste this URL into your browser: ")
-    print(authorization_url)
-    print(f"\nWaiting for authorization and callback to: {_REDIRECT_URI}")
+    click.echo("Paste this URL into your browser:\n")
+    click.echo(authorization_url)
+    click.echo(f"\nWaiting for authorization and callback to: {_REDIRECT_URI}\n")
 
     # Retrieves an authorization code by opening a socket to receive the
     # redirect request and parsing the query parameters set in the URL.
@@ -58,24 +131,9 @@ def auth(ctx: click.Context, secrets_file):
 
     # Pass the code back into the OAuth module to get a refresh token.
     flow.fetch_token(code=code)
+
     refresh_token = flow.credentials.refresh_token
-
-    click.echo(f"\nYour refresh token is: {refresh_token}\n")
-    click.echo("Updating config file to include refresh token")
-
-    with open(ctx.obj.config_file.path, "r") as f:
-        content = f.read()
-
-    with open(ctx.obj.config_file.path, "w") as f:
-        if "refresh_token" in content:
-            content = re.sub(
-                r"refresh_token = .*", f'refresh_token = "{refresh_token}"', content
-            )
-        else:
-            content = content + f'\nrefresh_token = "{refresh_token}"\n'
-        f.write(content)
-
-    click.echo("Done! You can now run adsctl commands.")
+    return refresh_token
 
 
 def get_authorization_code(passthrough_val):
@@ -96,6 +154,7 @@ def get_authorization_code(passthrough_val):
     # Parse the raw request to retrieve the URL query parameters.
     params = parse_raw_query_params(data)
 
+    message = ""
     try:
         if not params.get("code"):
             # If no code is present in the query params then there will be an