+++ src/adsctl/cli/edit/main.py
@@ -1,11 +1,16 @@
 import click
 
-from .campaign import campaign
+from adsctl.application import Application
+from adsctl.cli.edit.campaign import campaign
 
 
 @click.group()
-def edit():
-    pass
+@click.pass_obj
+def edit(app: Application):
+    try:
+        app.load_config()
+    except Exception as e:
+        click.echo(f"Error loading config: {e}")
 
 
 edit.add_command(campaign)